requirements.txt
selenium>=4.0.0
PyYAML>=6.0.2